+++ src/main/kotlin/io/kjson/JSONFun.kt
@@ -72,7 +72,7 @@ fun CharSequence.parseJSON(
 fun <T : Any> CharSequence.parseJSON(
     resultClass: KClass<T>,
     config: JSONConfig = JSONConfig.defaultConfig,
-): T? = JSONDeserializer.deserialize(resultClass, callParser(this.toString(), config), config)
+): T = JSONDeserializer.deserialize(resultClass, callParser(this.toString(), config), config)
 
 /**
  * Deserialize JSON from string ([CharSequence]) to the inferred [KType].
@@ -82,9 +82,9 @@ fun <T : Any> CharSequence.parseJSON(
  * @param   T       the target class
  * @return          the converted object
  */
-inline fun <reified T : Any> CharSequence.parseJSON(
+inline fun <reified T : Any?> CharSequence.parseJSON(
     config: JSONConfig = JSONConfig.defaultConfig,
-): T? = parseJSON(typeOf<T>(), config) as T?
+): T = parseJSON(typeOf<T>(), config) as T
 
 /**
  * Deserialize JSON from a [Reader] to a specified [KType].
@@ -116,7 +116,7 @@ fun Reader.parseJSON(
 fun <T : Any> Reader.parseJSON(
     resultClass: KClass<T>,
     config: JSONConfig = JSONConfig.defaultConfig,
-): T? {
+): T {
     val json = JSONStreamer(config.parseOptions).also {
         it.accept(this)
     }.result
@@ -133,7 +133,7 @@ fun <T : Any> Reader.parseJSON(
  */
 inline fun <reified T : Any> Reader.parseJSON(
     config: JSONConfig = JSONConfig.defaultConfig,
-): T? = parseJSON(typeOf<T>(), config) as T?
+): T = parseJSON(typeOf<T>(), config) as T
 
 /**
  * Invoke parser with parsing options from config.
@@ -217,8 +217,7 @@ fun Type.toKType(nullable: Boolean = false): KType = when (this) {
 /**
  * Deserialize a [JSONValue] to the nominated [KType].
  *
- * This function has been superseded by `fromJSONValue` and `fromJSONValueNullable`, and may be deprecated or removed in
- * future releases.
+ * This function has been superseded by `fromJSONValue`, and may be deprecated or removed in future releases.
  *
  * @receiver    the [JSONValue] (or `null`)
  * @param       resultType  the target [KType]
@@ -233,8 +232,7 @@ fun JSONValue?.deserialize(
 /**
  * Deserialize a [JSONValue] to the nominated [KClass].
  *
- * This function has been superseded by `fromJSONValue` and `fromJSONValueNullable`, and may be deprecated or removed in
- * future releases.
+ * This function has been superseded by `fromJSONValue`, and may be deprecated or removed in future releases.
  *
  * @receiver    the [JSONValue] (or `null`)
  * @param       resultClass the target [KClass]
@@ -245,13 +243,12 @@ fun JSONValue?.deserialize(
 fun <T : Any> JSONValue?.deserialize(
     resultClass: KClass<T>,
     config: JSONConfig = JSONConfig.defaultConfig,
-): T? = JSONDeserializer.deserialize(resultClass, this, config)
+): T = JSONDeserializer.deserialize(resultClass, this, config)
 
 /**
  * Deserialize a [JSONValue] to the inferred type.
  *
- * This function has been superseded by `fromJSONValue` and `fromJSONValueNullable`, and may be deprecated or removed in
- * future releases.
+ * This function has been superseded by `fromJSONValue`, and may be deprecated or removed in future releases.
  *
  * @receiver    the [JSONValue] (or `null`)
  * @param       config      an optional [JSONConfig] to customise the conversion
@@ -260,13 +257,12 @@ fun <T : Any> JSONValue?.deserialize(
  */
 inline fun <reified T : Any> JSONValue?.deserialize(
     config: JSONConfig = JSONConfig.defaultConfig,
-): T? = JSONDeserializer.deserialize(this, config)
+): T = JSONDeserializer.deserialize(this, config)
 
 /**
  * Deserialize a [JSONValue] to the nominated Java [Type].
  *
- * This function has been superseded by `fromJSONValue` and `fromJSONValueNullable`, and may be deprecated or removed in
- * future releases.
+ * This function has been superseded by `fromJSONValue`, and may be deprecated or removed in future releases.
  *
  * @receiver    the [JSONValue] (or `null`)
  * @param       javaType    the target [Type]
@@ -276,109 +272,61 @@ inline fun <reified T : Any> JSONValue?.deserialize(
 fun JSONValue?.deserialize(
     javaType: Type,
     config: JSONConfig = JSONConfig.defaultConfig,
-): Any? = JSONDeserializer.deserialize(javaType, this, config)
+): Any = JSONDeserializer.deserialize(javaType, this, config)
 
 /**
- * Deserialize a possibly null [JSONValue] to the nominated [KType].
+ * Deserialize a [JSONValue] to the nominated [KType].
  *
  * @receiver    the [JSONValue] (or `null`)
  * @param       resultType  the target [KType]
  * @param       config      an optional [JSONConfig] to customise the conversion
+ * @return      the converted value (may be `null` if the target type is nullable)
  * @throws      JSONKotlinException if the value can not be converted
  */
-fun JSONValue?.fromJSONValueNullable(
+fun JSONValue?.fromJSONValue(
     resultType: KType,
     config: JSONConfig = JSONConfig.defaultConfig,
 ): Any? = JSONDeserializer.deserialize(resultType, this, config)
 
 /**
- * Deserialize a possibly null [JSONValue] to the nominated [KClass].
- *
- * @receiver    the [JSONValue] (or `null`)
- * @param       resultClass the target [KClass]
- * @param       config      an optional [JSONConfig] to customise the conversion
- * @param       T           the target class
- * @throws      JSONKotlinException if the value can not be converted
- */
-fun <T : Any> JSONValue?.fromJSONValueNullable(
-    resultClass: KClass<T>,
-    config: JSONConfig = JSONConfig.defaultConfig,
-): T? = JSONDeserializer.deserialize(resultClass, this, config)
-
-/**
- * Deserialize a possibly null [JSONValue] to the inferred type.
- *
- * @receiver    the [JSONValue] (or `null`)
- * @param       config      an optional [JSONConfig] to customise the conversion
- * @param       T           the target class
- * @throws      JSONKotlinException if the value can not be converted
- */
-inline fun <reified T> JSONValue?.fromJSONValueNullable(
-    config: JSONConfig = JSONConfig.defaultConfig,
-): T? = JSONDeserializer.deserialize(this, config)
-
-/**
- * Deserialize a possibly null [JSONValue] to the nominated Java [Type].
- *
- * @receiver    the [JSONValue] (or `null`)
- * @param       javaType    the target [Type]
- * @param       config      an optional [JSONConfig] to customise the conversion
- * @throws      JSONKotlinException if the value can not be converted
- */
-fun JSONValue?.fromJSONValueNullable(
-    javaType: Type,
-    config: JSONConfig = JSONConfig.defaultConfig,
-): Any? = JSONDeserializer.deserialize(javaType, this, config)
-
-/**
- * Deserialize a [JSONValue] to the nominated [KType].
- *
- * @receiver    the [JSONValue]
- * @param       resultType  the target [KType]
- * @param       config      an optional [JSONConfig] to customise the conversion
- * @throws      JSONKotlinException if the value can not be converted
- */
-fun JSONValue.fromJSONValue(
-    resultType: KType,
-    config: JSONConfig = JSONConfig.defaultConfig,
-): Any = JSONDeserializer.deserializeNonNullable(resultType, this, config)
-
-/**
  * Deserialize a [JSONValue] to the nominated [KClass].
  *
- * @receiver    the [JSONValue]
+ * @receiver    the [JSONValue] (or `null`)
  * @param       resultClass the target [KClass]
  * @param       config      an optional [JSONConfig] to customise the conversion
  * @param       T           the target class
+ * @return      the converted value (may not be `null`)
  * @throws      JSONKotlinException if the value can not be converted
  */
 @Suppress("UNCHECKED_CAST")
-fun <T : Any> JSONValue.fromJSONValue(
+fun <T : Any> JSONValue?.fromJSONValue(
     resultClass: KClass<T>,
     config: JSONConfig = JSONConfig.defaultConfig,
-): T = JSONDeserializer.deserializeNonNullable(resultClass.starProjectedType, this, config) as T
+): T = JSONDeserializer.deserialize(resultClass.starProjectedType, this, config) as T
 
 /**
  * Deserialize a [JSONValue] to the inferred type.
  *
- * @receiver    the [JSONValue]
+ * @receiver    the [JSONValue] (or `null`)
  * @param       config      an optional [JSONConfig] to customise the conversion
  * @param       T           the target class
+ * @return      the converted value (may be `null` if the inferred type is nullable)
  * @throws      JSONKotlinException if the value can not be converted
  */
-inline fun <reified T : Any> JSONValue.fromJSONValue(
+inline fun <reified T : Any> JSONValue?.fromJSONValue(
     config: JSONConfig = JSONConfig.defaultConfig,
-): T = JSONDeserializer.deserializeNonNullable(typeOf<T>(), this, config) as T
+): T = JSONDeserializer.deserialize(typeOf<T>(), this, config) as T
 
 /**
  * Deserialize a [JSONValue] to the nominated Java [Type].
  *
- * @receiver    the [JSONValue]
+ * @receiver    the [JSONValue] (or `null`)
  * @param       javaType    the target [Type]
  * @param       config      an optional [JSONConfig] to customise the conversion
+ * @return      the converted value (may not be `null`)
  * @throws      JSONKotlinException if the value can not be converted
  */
-fun JSONValue.fromJSONValue(
+fun JSONValue?.fromJSONValue(
     javaType: Type,
     config: JSONConfig = JSONConfig.defaultConfig,
-): Any = JSONDeserializer.deserializeNonNullable(javaType.toKType(nullable = false), this, config)
+): Any = JSONDeserializer.deserialize(javaType.toKType(), this, config) as Any